requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
attrs==25.1.0
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
filelock==3.13.1
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.6.1
idna==3.10
Jinja2==3.1.4
kiwisolver==1.4.8
MarkupSafe==2.1.5
matplotlib==3.10.0
mpmath==1.3.0
multidict==6.1.0
networkx==3.3
numpy==2.1.2
packaging==24.2
pillow==11.0.0
propcache==0.3.0
pycocotools==2.0.8
pyparsing==3.2.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
six==1.17.0
sympy==1.13.1
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
torchmetrics==1.6.1
torchvision==0.21.0+cu124
tqdm==4.67.1
typing_extensions==4.12.2
yarl==1.18.3
